text_to_speech/requirements.txt
fastapi
uvicorn[standard]
pydantic
elevenlabs
python-dotenv
requests